requirements.txt
async-timeout==4.0.3
bleak==0.21.1
Pillow==10.1.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
typing_extensions==4.8.0
